js のインスタンスを使い回すよう修正

+++ src/main.rs
@@ -4,8 +4,8 @@ use std::sync::mpsc;
 use std::thread;
 
 fn main() {
-    let (input_tx, input_rx) = mpsc::channel::<Vec<f64>>();
-    let (output_tx, output_rx) = mpsc::channel::<Vec<f64>>();
+    let (input_tx, input_rx) = mpsc::channel::<Vec<f32>>();
+    let (output_tx, output_rx) = mpsc::channel::<Vec<f32>>();
     let th = thread::spawn(move || {
         let mut runtime: Box<dyn jsruntime::ScriptRuntime> = Box::new(jsruntime::JsRuntime::new());
         _ = (&mut *runtime).compile(
@@ -21,11 +21,11 @@ fn main() {
         "#,
         );
         let input = input_rx.recv().unwrap();
-        let mut output = vec![0f64; 4];
+        let mut output = vec![0f32; 4];
         _ = (&mut *runtime).process(&input, &mut output);
         let _ = output_tx.send(output);
     });
-    let input = vec![0f64, 1f64, 2f64, 5f64];
+    let input = vec![0f32, 1f32, 2f32, 5f32];
     let _ = input_tx.send(input);
     let _ = th.join();
     let output = output_rx.recv().unwrap();
